v1.0.6 Release

+++ apps/api/src/app/cards/cards.controller.ts
@@ -8,14 +8,15 @@ import {
   UseGuards,
   Body,
   Put,
-  Delete
+  Delete,
+  UnauthorizedException
 } from "@nestjs/common";
 import { Request as ExpressRequest } from "express";
 import { CardsService } from "./cards.service";
 import { UsersService } from "../users/users.service";
 import { AuthenticatedGuard } from "../auth/authenticated.guard";
 import { SetsService } from "../sets/sets.service";
-import { ApiResponse, CardIdParam, CreateCardDto, UpdateCardDto } from "@scholarsome/shared";
+import { ApiResponse, ApiResponseOptions, CardIdParam, CreateCardDto, UpdateCardDto } from "@scholarsome/shared";
 import { CreateCardGuard } from "./guards/create-card.guard";
 import { DeleteCardGuard } from "./guards/delete-card.guard";
 import { UpdateCardGuard } from "./guards/update-card.guard";
@@ -66,7 +67,7 @@ export class CardsController {
     }
 
     return {
-      status: "success",
+      status: ApiResponseOptions.Success,
       data: card
     };
   }
@@ -78,7 +79,9 @@ export class CardsController {
    */
   @UseGuards(AuthenticatedGuard, CreateCardGuard)
   @Post()
-  async createCard(@Body() body: CreateCardDto): Promise<ApiResponse<Card>> {
+  async createCard(@Body() body: CreateCardDto, @Request() req: ExpressRequest): Promise<ApiResponse<Card>> {
+    if (!(await this.setsService.verifySetOwnership(req, body.setId))) throw new UnauthorizedException();
+
     let media = [];
 
     const termScanned = await this.cardsService.scanAndUploadMedia(body.term, body.setId);
@@ -94,7 +97,7 @@ export class CardsController {
     }
 
     return {
-      status: "success",
+      status: ApiResponseOptions.Success,
       data: await this.cardsService.createCard({
         index: body.index,
         term: body.term,
@@ -124,9 +127,11 @@ export class CardsController {
    */
   @UseGuards(AuthenticatedGuard, UpdateCardGuard)
   @Put(":cardId")
-  async updateCard(@Param() params: CardIdParam, @Body() body: UpdateCardDto): Promise<ApiResponse<Card>> {
+  async updateCard(@Param() params: CardIdParam, @Body() body: UpdateCardDto, @Request() req: ExpressRequest): Promise<ApiResponse<Card>> {
     const card = await this.cardsService.card({ id: params.cardId });
     if (!card) throw new NotFoundException();
+
+    if (!(await this.setsService.verifySetOwnership(req, card.setId))) throw new UnauthorizedException();
 
     let media = [];
 
@@ -173,7 +178,7 @@ export class CardsController {
     }
 
     return {
-      status: "success",
+      status: ApiResponseOptions.Success,
       data: await this.cardsService.updateCard({
         where: {
           id: params.cardId
@@ -203,9 +208,17 @@ export class CardsController {
    */
   @UseGuards(AuthenticatedGuard, DeleteCardGuard)
   @Delete(":cardId")
-  async deleteCard(@Param() params: CardIdParam): Promise<ApiResponse<Card>> {
+  async deleteCard(@Param() params: CardIdParam, @Request() req: ExpressRequest): Promise<ApiResponse<Card>> {
+    const userCookie = this.usersService.getUserInfo(req);
+    if (!userCookie) throw new NotFoundException();
+
+    const card = await this.cardsService.card({ id: params.cardId });
+    if (!card) throw new NotFoundException();
+
+    if (card.set.authorId !== userCookie.id) throw new NotFoundException();
+
     return {
-      status: "success",
+      status: ApiResponseOptions.Success,
       data: await this.cardsService.deleteCard({
         id: params.cardId
       })
